+++ golang/lib/binding_constructors/binding_constructors.go
@@ -1,4 +1,0 @@
-package binding_constructors
-
-// The generated bindings don't come with constructors (leaving it up to the user to initialize all the fields), so we
-// add them so that our code is safer