+++ typescript/src/lib/kurtosis_context/kurtosis_context.ts
@@ -1,6 +1,5 @@
 import { EngineServiceClient } from "../../kurtosis_engine_rpc_api_bindings/engine_service_grpc_pb";
 import { LISTEN_PORT} from "../../kurtosis_engine_rpc_api_consts/kurtosis_engine_rpc_api_consts";
-import {EnclaveContext} from "../enclave_context/enclave_context";
 import * as grpc from "grpc";
 import { Result, err, ok, Err } from "neverthrow";
 import {newCreateEnclaveArgs, newDestroyEnclaveArgs, newStopEnclaveArgs} from "../constructor_calls";
@@ -11,12 +10,20 @@ import {
     EnclaveContainersStatusMap,
     EnclaveInfo, GetEnclavesResponse, StopEnclaveArgs
 } from "../../kurtosis_engine_rpc_api_bindings/engine_service_pb";
-import {ApiContainerContext} from "../api_container_context/api_container_context";
 import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
 import * as jspb from "google-protobuf";
+import { ApiContainerServiceClient, EnclaveContext, EnclaveID, KURTOSIS_API_VERSION } from "kurtosis-core-api-lib";
 
 const LOCAL_HOST_IP_ADDRESS_STR: string = "0.0.0.0";
 
+// TODO even the org-and-repo should come from Kurt Core
+const API_CONTAINER_IMAGE: string = "kurtosistech/kurtosis-core_api:" + KURTOSIS_API_VERSION;
+
+const SHOULD_PUBLISH_ALL_PORTS: boolean = true;
+
+const API_CONTAINER_LOG_LEVEL: string = "info";
+
+// Docs available at https://docs.kurtosistech.com/kurtosis-engine-api-lib/lib-documentation
 export class KurtosisContext {
     private readonly client: EngineServiceClient;
 
@@ -46,19 +53,19 @@ export class KurtosisContext {
         return ok(kurtosisContext);
     }
 
+    // Docs available at https://docs.kurtosistech.com/kurtosis-engine-api-lib/lib-documentation
     public async createEnclave(
         enclaveId: string,
-        apiContainerImage: string,
-        apiContainerLogLevel: string,
         isPartitioningEnabled: boolean,
-        shouldPublishPorts: boolean): Promise<Result<EnclaveContext, Error>> {
+    ): Promise<Result<EnclaveContext, Error>> {
 
         const args: CreateEnclaveArgs = newCreateEnclaveArgs(
             enclaveId,
-            apiContainerImage,
-            apiContainerLogLevel,
+            API_CONTAINER_IMAGE,
+            API_CONTAINER_LOG_LEVEL,
             isPartitioningEnabled,
-            shouldPublishPorts)
+            SHOULD_PUBLISH_ALL_PORTS,
+        );
 
         const createEnclavePromise: Promise<Result<CreateEnclaveResponse, Error>> = new Promise((resolve, _unusedReject) => {
             this.client.createEnclave(args, (error: grpc.ServiceError | null, response?: CreateEnclaveResponse) => {
@@ -93,8 +100,8 @@ export class KurtosisContext {
         return ok(newEnclaveContextResult.value);
     }
 
-    public async getEnclaves(): Promise<Result<Map<string, EnclaveContext>, Error>>{
-
+    // Docs available at https://docs.kurtosistech.com/kurtosis-engine-api-lib/lib-documentation
+    public async getEnclaveContext(enclaveId: EnclaveID): Promise<Result<EnclaveContext, Error>> {
         const emptyArg: google_protobuf_empty_pb.Empty = new google_protobuf_empty_pb.Empty()
 
         const getEnclavesPromise: Promise<Result<GetEnclavesResponse, Error>> = new Promise((resolve, _unusedReject) => {
@@ -110,20 +117,58 @@ export class KurtosisContext {
                 }
             })
         });
-
         const getEnclavesResult: Result<GetEnclavesResponse, Error> = await getEnclavesPromise;
         if (!getEnclavesResult.isOk()) {
             return err(getEnclavesResult.error)
         }
+        const getEnclavesResponse: GetEnclavesResponse = getEnclavesResult.value;
 
-        const response: GetEnclavesResponse = getEnclavesResult.value;
+        const allEnclaveInfo: jspb.Map<string, EnclaveInfo> = getEnclavesResponse.getEnclaveInfoMap()
+        const maybeEnclaveInfo: EnclaveInfo | undefined = allEnclaveInfo.get(enclaveId);
+        if (maybeEnclaveInfo === undefined) {
+            return err(new Error(`No enclave with ID '${enclaveId}' found`))
+        }
+        const enclaveInfo: EnclaveInfo = maybeEnclaveInfo;
 
-        const enclavesMap: Map<string, EnclaveContext> = this.newEnclaveContextMapFromEnclaveInfoMap(response.getEnclaveInfoMap())
-
-        return ok(enclavesMap)
+        const newEnclaveCtxResult: Result<EnclaveContext, Error> = this.newEnclaveContextFromEnclaveInfo(enclaveInfo);
+        if (newEnclaveCtxResult.isErr()) {
+            return err(newEnclaveCtxResult.error);
+        }
+        return ok(newEnclaveCtxResult.value);
     }
 
-    public async stopEnclave(enclaveId: string): Promise<Result<null, Error>> {
+    // Docs available at https://docs.kurtosistech.com/kurtosis-engine-api-lib/lib-documentation
+    public async getEnclaves(): Promise<Result<Set<EnclaveID>, Error>>{
+        const emptyArg: google_protobuf_empty_pb.Empty = new google_protobuf_empty_pb.Empty()
+
+        const getEnclavesPromise: Promise<Result<GetEnclavesResponse, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.getEnclaves(emptyArg, (error: grpc.ServiceError | null, response?: GetEnclavesResponse) => {
+                if (error === null) {
+                    if (!response) {
+                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
+                    } else {
+                        resolve(ok(response!));
+                    }
+                } else {
+                    resolve(err(error));
+                }
+            })
+        });
+        const getEnclavesResult: Result<GetEnclavesResponse, Error> = await getEnclavesPromise;
+        if (!getEnclavesResult.isOk()) {
+            return err(getEnclavesResult.error)
+        }
+        const getEnclavesResponse: GetEnclavesResponse = getEnclavesResult.value;
+
+        const result: Set<EnclaveID> = new Set();
+        for (let enclaveId of getEnclavesResponse.getEnclaveInfoMap().keys()) {
+            result.add(enclaveId);
+        }
+        return ok(result);
+    }
+
+    // Docs available at https://docs.kurtosistech.com/kurtosis-engine-api-lib/lib-documentation
+    public async stopEnclave(enclaveId: EnclaveID): Promise<Result<null, Error>> {
         const args: StopEnclaveArgs = newStopEnclaveArgs(enclaveId)
 
         const stopEnclavePromise: Promise<Result<null, Error>> = new Promise((resolve, _unusedReject) => {
@@ -143,7 +188,8 @@ export class KurtosisContext {
         return ok(null);
     }
 
-    public async destroyEnclave(enclaveId: string): Promise<Result<null, Error>> {
+    // Docs available at https://docs.kurtosistech.com/kurtosis-engine-api-lib/lib-documentation
+    public async destroyEnclave(enclaveId: EnclaveID): Promise<Result<null, Error>> {
         const args: DestroyEnclaveArgs = newDestroyEnclaveArgs(enclaveId);
 
         const destroyEnclavePromise: Promise<Result<null, Error>> = new Promise((resolve, _unusedReject) => {
@@ -166,24 +212,6 @@ export class KurtosisContext {
     // ====================================================================================================
     //                                       Private helper functions
     // ====================================================================================================
-    private newEnclaveContextMapFromEnclaveInfoMap(enclaveInfoMap: jspb.Map<string, EnclaveInfo>): Map<string, EnclaveContext> {
-
-        const enclaveContextMap: Map<string, EnclaveContext> = new Map<string, EnclaveContext>()
-
-        for (const [key, value] of enclaveInfoMap.entries()) {
-            const newEnclaveContextResult: Result<EnclaveContext, Error> = this.newEnclaveContextFromEnclaveInfo(value);
-            if (newEnclaveContextResult.isErr()) {
-                // TODO This is really nasty - we skip enclaves that we can't create, and the only reason we wouldn't be able to create
-                //  enclaves is because of stopped containers. Basically, we should move the API container into the engine-server, and
-                //  make it impossible for enclaves to be in an invalid state at all
-                continue;
-            }
-            enclaveContextMap.set(key, newEnclaveContextResult.value);
-        }
-
-        return enclaveContextMap;
-    }
-
     private newEnclaveContextFromEnclaveInfo(enclaveInfo: EnclaveInfo): Result<EnclaveContext, Error> {
         const enclaveContainersStatus = enclaveInfo.getContainersStatus()
         if (enclaveContainersStatus !== EnclaveContainersStatus.ENCLAVECONTAINERSSTATUS_RUNNING) {
@@ -204,21 +232,26 @@ export class KurtosisContext {
             return err(new Error(`API container was listed as running, but no API container host machine info exists`))
         }
 
-        const apiContainerContext: ApiContainerContext = new ApiContainerContext(
-            apiContainerInfo.getContainerId(),
-            apiContainerInfo.getIpInsideEnclave(),
-            apiContainerInfo.getPortInsideEnclave(),
-            apiContainerHostMachineInfo.getIpOnHostMachine(),
-            apiContainerHostMachineInfo.getPortOnHostMachine()
-        )
+        const apiContainerHostMachineUrl: string = `${apiContainerHostMachineInfo.getIpOnHostMachine()}:${apiContainerHostMachineInfo.getPortOnHostMachine()}`
 
-        const enclaveContext: EnclaveContext = new EnclaveContext(
+        let apiContainerClient: ApiContainerServiceClient;
+        // TODO SECURITY: Use HTTPS!
+        try {
+            apiContainerClient = new ApiContainerServiceClient(apiContainerHostMachineUrl, grpc.credentials.createInsecure());
+        } catch(exception) {
+            if (exception instanceof Error) {
+                return err(exception);
+            }
+            return err(new Error(
+                "An unknown exception value was thrown during creation of the API container client that wasn't an error: " + exception
+            ));
+        }
+
+        const result: EnclaveContext = new EnclaveContext(
+            apiContainerClient,
             enclaveInfo.getEnclaveId(),
-            enclaveInfo.getNetworkId(),
-            enclaveInfo.getNetworkCidr(),
-            apiContainerContext,
+            enclaveInfo.getEnclaveDataDirpathOnHostMachine(),
         )
-
-        return ok(enclaveContext);
+        return ok(result);
     }
 }