+++ docs/changelog.md
@@ -1,5 +1,19 @@
 # TBD
 
+# 0.10.0
+### Features
+* Added `KurtosisContext.getEnclaveContext` function
+* Added docs for `KurtosisContext`
+
+### Changes
+* `KurtosisContext.getEnclaves` now simply returns a set of enclave IDs
+
+### Breaking Changes
+* `NetworkContext` has been renamed to `EnclaveContext`
+* `KurtosisContext` enclave manipulation returns & accepts `EnclaveContext` objects from `kurtosis-client`
+* `KurtosisContext.getEnclaves` returns a set of enclave IDs
+* The `EnclaveInfo` object has an additional field, `enclaveDataDirpathOnHostMachine`
+
 # 0.9.0
 ### Removals
 * Remove the API container endpoints that were added in 0.8.0, because we're backing out of combining the APIC with the engine server